requirements.txt
fastapi
uvicorn
python-multipart
PyJWT
sqlalchemy
psycopg2-binary
pypdf
passlib
bcrypt==4.0.1